fix: Convert all CRLFs to LFs
Motivation: Bale's OCD Inventory Control used CRLFs for line endings,
which we inherited when we forked the project. However, many tools
expect and use LF endings. This includes GitHub's web editor UI, which
automatically converts all files to use LF endings. This is a problem
because Git treats a LF -> CRLF conversion as a complete change and
creates a large diff log that makes it hard to track changes (e.g. git
blame). Maintaining a project with a mix of LFs and CRLFs is also
tricky.

To address these problems, this commit converts ALL files to use LF
endings.

Since nearly all contemporary text editors (VS Code, Notepad++, and
Notepad in Windows 10) support LF endings, this shouldn't be a problem
nowadays.

See this comment, which motivated this change:
https://github.com/Loathing-Associates-Scripting-Society/OCD-Cleanup/pull/13#issuecomment-756989865