fix: Show correct expected price when using mall multi
When sending MALL items to a mall multi, OCD-Cleanup previously
displayed an incorrect "Sale price for this line:" message, usually with
a price of zero. Fix this so that the script always shows the sum of
_expected_ sale prices.

Note: The _expected_ sale price always uses the mall minimum (affected
by the user-defined minimum OCD price), even if `BaleOCD_Pricing` is
not set to "auto".