refactor(manager): Simplify keyboard scroll logic
Instead of listening to keyDown events and manually calling `scrollTo()`
and `scrollBy()`, set a `tabIndex` on the react-window container element
and rely on the browser's default scrolling logic.

This is the "strategy" used by the <Table> component in
react-virtualized. It's much more "natural" than the method suggested by
Brian Vaughn, the author of react-window:
https://github.com/bvaughn/react-window/issues/46#issuecomment-416073707