fix(manager): Create reproducible builds
Vite currently has a bug that causes it to generate different bundles on
Windows and Linux. This is caused by a fault in @rollup/plugin-commonjs
which is bundled into Vite.

We directly fix the fault in our `node_modules/` using patch-package.
Let's hope the Rollup devs accept my PR and the fix is integrated into
Vite.

More links:

- Bug reports:
    - https://github.com/vitejs/vite/issues/2623
    - https://github.com/rollup/plugins/issues/923
- Related pull requests:
    - https://github.com/rollup/plugins/pull/924