fix(manager): show correct item name in toast msg
When changing the action associated with an item, the toast message
showed an incorrect item name or "undefined". This was because the code
was incorrectly retrieving an item from the `data.items` array by using
`itemId` as the index, rather than performing a lookup via `.find()`.
This commit fixes this.