fix(manager): show 'saving' toast only when actually making requests
Previously, the "Saving..." toast message was shown immediately when the
user clicked the "Save" button. This is because the toast message is
tied to the `isSaving` variable, which is set while `saveConfig()` is
running. This caused a slight lag when it launched the "Do you want to
copy/rename the files?" modal dialog: the toast message would appear,
then be immediately overridden by the dialog.

Fix this by moving most of the synchronous checks (including launching
the modal) into a separate `saveConfigWithCheck()` function. Now,
`saveConfig()` is only called (and the toast message shown) after all
checks are made and the modal is closed.